feat: improve loading error message
sometimes the caching service worker causes the following fatal error when loading the page:

A ServiceWorker passed a promise to FetchEvent.respondWith() that rejected with 'TypeError'

i've been unable to figure out what causes it/reproduce it when running the site locally, so just telling users how to potentially fix it instead.

thought about removing the service worker entirely but it's pretty essential for a nice user experience.